# EmprestimosOpenAPI.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "EmprestimosOpenAPI | v1",
    "version": "1.0.0"
  },
  "paths": {
    "/v1/emprestimos": {
      "post": {
        "tags": [
          "Empréstimos V1"
        ],
        "summary": "Cria um novo empréstimo",
        "description": "Cria um novo empréstimo com os dados fornecidos no corpo da requisição.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Emprestimo"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Emprestimo"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Empréstimos V1"
        ],
        "summary": "Lista todos os empréstimos",
        "description": "Retorna a lista completa de empréstimos cadastrados no sistema.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Emprestimo"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/emprestimos/{id}": {
      "get": {
        "tags": [
          "Empréstimos V1"
        ],
        "summary": "Consulta empréstimo por ID",
        "description": "Retorna um empréstimo específico identificado pelo GUID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Emprestimo"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "Empréstimos V1"
        ],
        "summary": "Atualiza empréstimo",
        "description": "Atualiza completamente os dados de um empréstimo existente.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Emprestimo"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Emprestimo"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "patch": {
        "tags": [
          "Empréstimos V1"
        ],
        "summary": "Atualiza parcialmente",
        "description": "Atualiza parcialmente os dados de um empréstimo, como valor ou prazo.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AtualizacaoEmprestimoDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Emprestimo"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "Empréstimos V1"
        ],
        "summary": "Remove empréstimo",
        "description": "Remove um empréstimo do sistema pelo seu identificador único.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AtualizacaoEmprestimoDto": {
        "required": [
          "valor",
          "prazoMeses"
        ],
        "type": "object",
        "properties": {
          "valor": {
            "maximum": 100000,
            "minimum": 100,
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "prazoMeses": {
            "maximum": 120,
            "minimum": 2,
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        }
      },
      "Emprestimo": {
        "required": [
          "cliente",
          "valor"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "cliente": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "valor": {
            "maximum": 100000,
            "minimum": 100,
            "type": "number",
            "format": "double"
          },
          "prazoMeses": {
            "maximum": 120,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "taxaJurosMensal": {
            "type": "number",
            "format": "double"
          },
          "dataContrato": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "ValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Empréstimos V1"
    }
  ]
}